refactor(p2-shim): extract HTTP future browser fixture
